+++ railties/test/generators/scaffold_controller_generator_test.rb
@@ -185,4 +185,15 @@ class ScaffoldControllerGeneratorTest < Rails::Generators::TestCase
       assert_match(/2 runs, 2 assertions, 0 failures, 0 errors/, `bundle exec rake test 2>&1`)
     end
   end
+
+  def test_controller_tests_pass_by_default_inside_full_engine
+    Dir.chdir(destination_root) { `bundle exec rails plugin new bukkits --full` }
+
+    engine_path = File.join(destination_root, "bukkits")
+
+    Dir.chdir(engine_path) do
+      quietly { `bin/rails g controller dashboard foo` }
+      assert_match(/2 runs, 2 assertions, 0 failures, 0 errors/, `bundle exec rake test 2>&1`)
+    end
+  end
 end